Update readme
AOFX no longer depends on AGS. It never did really depend on AGS, but
there was a compile dependency in AMD_LIB on AGS even if code that
relied on AMD_LIB did not use the AGS-related functionality. That
compile dependency has since been removed in cases where the AGS-related
functionality is not used.